// web/src/client/models/ModelProviderUpdate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ModelProviderUpdate = {
    provider_name?: (string | null);
    base_url?: (string | null);
    api_key?: (string | null);
    icon?: (string | null);
    description?: (string | null);
};